refactor: featureize db package and harden server boundaries
