+++ apps/server/src/lib/ttl-cache.ts
@@ -1,0 +1,61 @@
+type CacheEntry<T> = {
+  expiresAt: number;
+  value: T;
+};
+
+export class TtlCache<K, V> {
+  private readonly entries = new Map<K, CacheEntry<V>>();
+
+  constructor(
+    private readonly ttlMs: number,
+    private readonly maxSize: number,
+  ) {}
+
+  get(key: K): V | undefined {
+    const entry = this.entries.get(key);
+
+    if (!entry) {
+      return undefined;
+    }
+
+    if (entry.expiresAt <= Date.now()) {
+      this.entries.delete(key);
+      return undefined;
+    }
+
+    this.entries.delete(key);
+    this.entries.set(key, entry);
+    return entry.value;
+  }
+
+  set(key: K, value: V): void {
+    this.pruneExpired();
+
+    if (this.entries.has(key)) {
+      this.entries.delete(key);
+    }
+
+    this.entries.set(key, {
+      expiresAt: Date.now() + this.ttlMs,
+      value,
+    });
+
+    while (this.entries.size > this.maxSize) {
+      const oldestKey = this.entries.keys().next().value;
+      if (oldestKey === undefined) {
+        break;
+      }
+      this.entries.delete(oldestKey);
+    }
+  }
+
+  private pruneExpired(): void {
+    const now = Date.now();
+
+    for (const [key, entry] of this.entries) {
+      if (entry.expiresAt <= now) {
+        this.entries.delete(key);
+      }
+    }
+  }
+}
