+++ apps/server/src/middleware/auth.middleware.ts
@@ -2,7 +2,7 @@ import { Elysia } from "elysia";
 import { auth } from "@anilog/auth";
 import { unauthorizedError } from "../lib/api-error";
 
-export const authPlugin = (app: Elysia) =>
+export const authMiddleware = (app: Elysia) =>
   app.derive(async ({ request }) => {
     const session = await auth.api.getSession({ headers: request.headers });
     const userId = session?.user?.id;
