+++ packages/db/src/index.ts
@@ -1,7 +1,5 @@
 export * from "./client";
-export * from "./queries/admin.query";
-export * from "./queries/anime.query";
-export * from "./queries/library.query";
-export * from "./queries/shared/follow-counts";
-export * from "./queries/shared/user-summary";
-export * from "./queries/users.query";
+export * from "./features/admin/admin.query";
+export * from "./features/anime/anime.query";
+export * from "./features/library/library.query";
+export * from "./features/users/users.query";
