+++ packages/db/src/schema/index.ts
@@ -1,2 +1,5 @@
-export * from "./auth";
-export * from "./anilog";
+export * from "../features/admin/admin.schema";
+export * from "../features/anime/anime.schema";
+export * from "../features/auth/auth.schema";
+export * from "../features/library/library.schema";
+export * from "../features/users/users.schema";
